added fade to question card components
