+++ package.json
@@ -1,0 +1,22 @@
+{
+  "name": "amazon_warfare",
+  "version": "1.0.0",
+  "description": "- Anthony Baietto - Jared Clabough - Sarah Flanagan - Dan Filler - Srini Srinivas",
+  "main": "server.js",
+  "scripts": {
+    "start-server": "nodemon  server.js",
+    "start-ui": "npm run start --prefix ui",
+    "start-dev": "concurrently \"npm run start-server\" \"npm run start-ui\"",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "body-parser": "^1.19.0",
+    "concurrently": "^5.1.0",
+    "express": "^4.17.1"
+  },
+  "devDependencies": {
+    "nodemon": "^2.0.2"
+  }
+}
