+++ ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mockup",
+  "name": "ui",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -11,6 +11,7 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-scripts": "3.3.0",
+    "react-transition-group": "^1.2.1",
     "reactstrap": "^8.4.0"
   },
   "scripts": {
