+++ ui/src/HigherOrderComponents/Fade.css
@@ -1,0 +1,14 @@
+.fade-appear {
+    opacity: 0.01;
+}
+.fade-appear.fade-appear-active {
+    opacity: 1;
+    transition: opacity 500ms ease-in;
+}
+.fade-leave {
+    opacity: 1;
+}
+.fade-leave.fade-leave-active {
+    opacity: 0.01;
+    transition: opacity 500ms ease-in;
+}
