+++ ui/src/components/Inputs/ButtonList/Button.scss
@@ -10,10 +10,11 @@
 }
 
 .button:hover {
+  transition-duration: .2s;
   background: $orange-red;
   color: $white;
 }
 
 .button:active {
-  background: $dark-charcoal;
+  background: $red;
 }
