+++ ui/src/components/NavBar/NavBar.scss
