+++ ui/src/components/NavBar/NavBarLogo.js
