+++ ui/src/components/QuestionCard/QuestionCard.scss
@@ -4,12 +4,13 @@
   background: $red;
   background: $red;
   color: $charcoal;
-  border: 2px solid white;
+  border: 1px solid $red;
 
   width: 50%;
   border-radius: 10px;
   margin: 5% auto;
 }
+
 
 .question-text {
   font-size: 30px;
